+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * ArcEllipse3D: discretization_points
 * FullArc3D: linesegment_intersections
 * Line: sort_points_along_line
+* Line2D: point_belongs
 * ArcEllipse2D: length, point_belongs, abscissa, bounding_rectangle, straight_line_area, discretization_points, reverse
 * ArcEllipse2D: straight_line_area
 * Line3D: skew_to: Verifies if two Line3D are skew
@@ -39,7 +40,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Ellipse2D.frame_mapping()
 * Line2D.frame_mapping()
 
-
 ### Fixed
 
 * BsplineCurve: abscissa (use different start point between 0 and length)
@@ -51,10 +51,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Step: from stream (use BinaryFile instead of io.BinaryIO)
 * Contour: is_overlapping (consider intersecting_points is empty)
 * LineSegment2D: to_wire (use discretization_points instead of discretise)
+* Line2D: line_intersections
+* BSplineCurve2D: line_intersections
 * ArcEllipse2D: to_3d
 * Fix boolean operations when faces are 100% coincident
 * Fix some to_step methods from edges.py and faces.py
-
 
 ### Performance improvements
 
@@ -82,6 +83,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * LineSegment3D: linesegment_intersections
 * Contour: is_overlapping
 * LineSegment2D: to_wire
+* Line2D: point_belongs
+* BSplineCurve2D: line_intersections
+* LineSegment2D: line_intersections
 * ArcEllipse3D: discretization_points
 * FullArc3D: linesegment_intersections
 * Line2D: sort_points_along_line